Explicit narrowing

+++ components/ble_adv_light/ble_adv_light.h
@@ -68,8 +68,8 @@ class ZhiJiaLight : public BleAdvLight
   void update_channels(uint8_t cold, uint8_t warm) override {
     ESP_LOGD("zhijia", "ZhiJiaLight::update_channels called (cold: %d, warm: %d)!", cold, warm);
     ESP_LOGD("zhijia", "Calling send_packet(0x%02x, %d)", CMD_CCT(), (uint8_t) (255 * ((float) warm / (cold + warm))));
-    uint8_t cct_args[1] = {(255 * ((float) warm / (cold + warm)))};
-    uint8_t dim_args[1] = {(cold + warm > 255 ? 255 : cold + warm)};
+    uint8_t cct_args[1] = {(uint8_t) (255 * ((float) warm / (cold + warm)))};
+    uint8_t dim_args[1] = {(uint8_t) (cold + warm > 255 ? 255 : cold + warm)};
     send_packet(CMD_CCT(), cct_args);
     send_packet(CMD_DIM(), dim_args);
   };
